deployment/requirements.txt
numpy==1.26.4
streamlit
Pillow
scikit-learn
matplotlib
seaborn